+++ Services/CommandBus.php
@@ -5,11 +5,15 @@ declare(strict_types=1);
 namespace RevisionTen\CQRS\Services;
 
 use Ramsey\Uuid\Uuid;
+use RevisionTen\CQRS\Exception\CommandValidationException;
 use RevisionTen\CQRS\Exception\InterfaceException;
 use RevisionTen\CQRS\Interfaces\AggregateInterface;
 use RevisionTen\CQRS\Interfaces\CommandInterface;
+use RevisionTen\CQRS\Interfaces\EventInterface;
 use RevisionTen\CQRS\Interfaces\HandlerInterface;
 use RevisionTen\CQRS\Message\Message;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 
 class CommandBus
 {
@@ -29,6 +33,11 @@ class CommandBus
     public $aggregateFactory;
 
     /**
+     * @var ContainerInterface
+     */
+    private $container;
+
+    /**
      * @var array
      */
     private $aggregates = [];
@@ -36,15 +45,17 @@ class CommandBus
     /**
      * CommandBus constructor.
      *
-     * @param EventBus         $eventBus
-     * @param MessageBus       $messageBus
-     * @param AggregateFactory $aggregateFactory
+     * @param \RevisionTen\CQRS\Services\EventBus                       $eventBus
+     * @param \RevisionTen\CQRS\Services\MessageBus                     $messageBus
+     * @param \RevisionTen\CQRS\Services\AggregateFactory               $aggregateFactory
+     * @param \Symfony\Component\DependencyInjection\ContainerInterface $container
      */
-    public function __construct(EventBus $eventBus, MessageBus $messageBus, AggregateFactory $aggregateFactory)
+    public function __construct(EventBus $eventBus, MessageBus $messageBus, AggregateFactory $aggregateFactory, ContainerInterface $container)
     {
         $this->eventBus = $eventBus;
         $this->messageBus = $messageBus;
         $this->aggregateFactory = $aggregateFactory;
+        $this->container = $container;
     }
 
     /**
@@ -76,6 +87,92 @@ class CommandBus
     }
 
     /**
+     * This performs the following actions:
+     * Get the Aggregate.
+     * Check if the Command is valid.
+     * Create an Event for the Command and push it onto the Aggregates pending Events.
+     *
+     * @param \RevisionTen\CQRS\Interfaces\CommandInterface $command
+     * @param \RevisionTen\CQRS\Interfaces\HandlerInterface $handler
+     *
+     * @throws \Exception
+     */
+    private function handleCommand(CommandInterface $command, HandlerInterface $handler): void
+    {
+        $aggregateClass = $command::getAggregateClass();
+        $aggregateUuid = $command->getAggregateUuid();
+        $user = $command->getUser();
+
+        // Get Aggregate.
+        $aggregate = $this->aggregateFactory->build($aggregateUuid, $aggregateClass, null, $user);
+
+        // Check if commands input is valid.
+        try {
+            $validCommand = $handler->validateCommand($command, $aggregate);
+        } catch (CommandValidationException $commandValidationException) {
+
+            $validCommand = false;
+
+            $this->messageBus->dispatch(new Message(
+                $commandValidationException->getMessage(),
+                $commandValidationException->getCode(),
+                $commandValidationException->command->getUuid(),
+                $commandValidationException->command->getAggregateUuid(),
+                $commandValidationException
+            ));
+        }
+
+        // Check if the Aggregate and target Version matches.
+        $versionMatches = $aggregate->getVersion() === $command->getOnVersion();
+
+        if (!$versionMatches) {
+            // Version does not match.
+            $this->messageBus->dispatch(new Message(
+                'Aggregate target version is outdated or does not exist',
+                CODE_CONFLICT,
+                $command->getUuid(),
+                $aggregateUuid,
+                null
+            ));
+        } elseif ($validCommand) {
+            try {
+                /**
+                 * Create Event for Command.
+                 *
+                 * @var EventInterface $event
+                 */
+                $event = $handler->createEvent($command);
+
+                if ($event instanceof EventInterface) {
+                    // Apply Event to Aggregate.
+                    $aggregate = $this->aggregateFactory->apply($aggregate, $event);
+
+                    $this->aggregates[] = $aggregate;
+                } else {
+                    throw new InterfaceException(\get_class($event).' must implement '.EventInterface::class);
+                }
+            } catch (InterfaceException $e) {
+                $this->messageBus->dispatch(new Message(
+                    $e->getMessage(),
+                    $e->getCode(),
+                    $command->getUuid(),
+                    $aggregateUuid,
+                    $e
+                ));
+            }
+        } else {
+            // Handle invalid command.
+            $this->messageBus->dispatch(new Message(
+                'Invalid Command',
+                CODE_BAD_REQUEST,
+                $command->getUuid(),
+                $aggregateUuid,
+                null
+            ));
+        }
+    }
+
+    /**
      * This function is used to dispatch a provided Command.
      *
      * @param \RevisionTen\CQRS\Interfaces\CommandInterface $command
@@ -97,11 +194,18 @@ class CommandBus
                  * @var HandlerInterface $handler
                  */
                 $handlerClass = $command::getHandlerClass();
-                $handler = new $handlerClass($this->messageBus, $this->aggregateFactory);
+
+                // Try to get the handler as a service or instantiate it.
+                try {
+                    $handler = $this->container->get($handlerClass);
+                } catch (ServiceNotFoundException $e) {
+                    $handler = new $handlerClass();
+                }
 
                 if ($handler instanceof HandlerInterface) {
-                    // Invoke Handler.
-                    $handler($command, $this->aggregates);
+
+                    // Use the command handler to validate the command and produce events.
+                    $this->handleCommand($command, $handler);
 
                     // Check for events on aggregates, pass them to EventBus#publish()
                     $events = [];